+++ .formatter.exs
@@ -1,5 +1,7 @@
 locals_without_parens = [
   defvocab: 2,
+  defnamespace: 2,
+  defnamespace: 3,
   def_facet_constraint: 2,
   def_applicable_facet: 1,
   bgp: 1,